All global states, together with the context wrapper, were written to the "AppStateProvider.jsx" component and moved to the "wrapper" folder, which is located in the "context" folder. To use the application context, the "useAppState.js" hook was created and placed in the "hook" folder, which in turn is located in the "context" folder. The state associated with the "TotalPrice" component was transferred to the component itself, and the calculations for calculating the final price were moved to a separate hook "useCalcutateTotalPice.js" which is located in the "hooks" folder, which in turn is located in the "src" folder. The wrapper of the application itself was moved to a separate component "Wrapper.jsx" which is located in the folder of the same name, which in turn is located in the "src" folder. Also, the "TotalPrice" component uses the context data with the "useAppState.js" hook call, sets the "className" using the "classNames" utility, and performs the total price calculation with the "useCalcutateTotalPice.js" hook.
