+++ src/components/wrapper/Wrapper.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import classNames from 'classnames';
+import useAppState from '../../context/hook/useAppState';
+
+function Wrapper(props) {
+  const {theme} = useAppState();
+  const className = classNames('app', theme);
+
+  return (
+    <div className={className}>
+      <div className='content'>{props.children}</div>
+    </div>
+  );
+}
+
+export default Wrapper;
