+++ src/context/hook/useAppState.js
@@ -1,0 +1,8 @@
+import {useContext} from 'react';
+import {ContextApp} from '../ContextApp';
+
+function useAppState() {
+  return useContext(ContextApp);
+}
+
+export default useAppState;
